requirements.txt
Flask==0.12.2
cf-deployment-tracker==1.0.4
cloudant==2.4.0
six==1.10.0
mysql-connector==2.1.6
